requirements.txt
boto3
kaggle
dbt-snowflake
apache-airflow==2.10.4
apache-airflow-providers-snowflake
apache-airflow-providers-amazon
awscli
